AppServer_SIGUI/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
gsconfig-py3==1.0.8
idna==3.3
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
